services/users/requirements.txt
flask==2.1.1
Flask-Admin==1.6.0
flask-bcrypt==1.0.1
flask-cors==3.0.10
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
gunicorn==20.1.0
psycopg2-binary==2.9.3
pyjwt==2.3.0
redis
kafka-python
pymongo
flask-mongoengine
